backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
# passlib[bcrypt] is broken down into two specific versions to ensure compatibility
passlib==1.7.4
bcrypt==3.2.2
psycopg2-binary
redis
celery
opencv-python
scikit-image
Pillow
python-multipart
google-generativeai